requirements.txt
# Core AI & LLM - Let pip resolve versions
google-generativeai
langgraph
langchain-google-genai

# Database
psycopg2-binary
sqlalchemy

# Configuration & Validation
pydantic
pydantic-settings
python-dotenv

# CLI & Output
rich
prompt-toolkit

# Utilities
typing-extensions